requirements.txt
streamlit_authenticator
streamlit
deta
python-dotenv
PyPDF2
Pillow
PyPDF2[image]
